+++ hubdle-gradle-plugin/main/kotlin/com/javiersc/hubdle/extensions/config/binary/compatibility/validator/HubdleBinaryCompatibilityValidatorExtension.kt
@@ -12,11 +12,13 @@ import org.gradle.api.model.ObjectFactory
 import org.gradle.kotlin.dsl.newInstance
 
 @HubdleDslMarker
-public open class BinaryCompatibilityValidatorExtension
+public open class HubdleBinaryCompatibilityValidatorExtension
 @Inject
 constructor(
     objects: ObjectFactory,
-) : EnableableOptions, RawConfigOptions<BinaryCompatibilityValidatorExtension.RawConfigExtension> {
+) :
+    EnableableOptions,
+    RawConfigOptions<HubdleBinaryCompatibilityValidatorExtension.RawConfigExtension> {
 
     override var Project.isEnabled: Boolean
         get() = hubdleState.config.binaryCompatibilityValidator.isEnabled