+++ hubdle-gradle-plugin/main/kotlin/com/javiersc/hubdle/extensions/config/versioning/HubdleVersioningExtension.kt
@@ -1,26 +1,30 @@
-package com.javiersc.hubdle.extensions.config.coverage
+package com.javiersc.hubdle.extensions.config.versioning
 
 import com.javiersc.hubdle.extensions.HubdleDslMarker
 import com.javiersc.hubdle.extensions._internal.state.hubdleState
 import com.javiersc.hubdle.extensions.options.EnableableOptions
 import com.javiersc.hubdle.extensions.options.RawConfigOptions
+import com.javiersc.semver.gradle.plugin.SemverExtension
 import javax.inject.Inject
-import kotlinx.kover.api.KoverExtension
 import org.gradle.api.Action
 import org.gradle.api.Project
 import org.gradle.api.model.ObjectFactory
 import org.gradle.kotlin.dsl.newInstance
 
 @HubdleDslMarker
-public open class CoverageExtension
+public abstract class HubdleVersioningExtension
 @Inject
 constructor(
     objects: ObjectFactory,
-) : EnableableOptions, RawConfigOptions<CoverageExtension.RawConfigExtension> {
+) : EnableableOptions, RawConfigOptions<HubdleVersioningExtension.RawConfigExtension> {
 
     override var Project.isEnabled: Boolean
-        get() = hubdleState.config.coverage.isEnabled
-        set(value) = hubdleState.config.coverage.run { isEnabled = value }
+        get() = hubdleState.config.versioning.isEnabled
+        set(value) = hubdleState.config.versioning.run { isEnabled = value }
+
+    public var Project.tagPrefix: String
+        get() = hubdleState.config.versioning.tagPrefix
+        set(value) = hubdleState.config.versioning.run { tagPrefix = value }
 
     override val rawConfig: RawConfigExtension = objects.newInstance()
 
@@ -32,8 +36,8 @@ constructor(
     @HubdleDslMarker
     public open class RawConfigExtension {
 
-        public fun Project.kover(action: Action<KoverExtension>) {
-            hubdleState.config.coverage.rawConfig.kover = action
+        public fun Project.semver(action: Action<SemverExtension>) {
+            hubdleState.config.versioning.rawConfig.semver = action
         }
     }
 }